Split mobile launch timing helpers
